requirements.txt
Flask==3.1.0
openai==1.55.3
python-dotenv==1.0.1